requirements.txt
asgiref==3.8.1
coverage==7.6.1
Django==5.1
djangorestframework==3.15.2
mysqlclient==2.2.4
PyJWT==2.9.0
python-dotenv==1.0.1
sqlparse==0.5.1
tzdata==2024.1